+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+- Fix bug with Open3 not being required
+
 ## [0.2.0] - 2020-11-21
 
 ### Changed